Added Unity Integration and Server Component
Added a C# server integration Visual Studio project that allows you to
open ports and accept communications using the .NET protocol. Will most
likely change this to C++ in the near future.

Added a Unity Scene with the C# scripts necessary to make a TCP
connection to the C# server. (Will be updated when server changed to
C++)

Most of this work was done offline over time and am only now uploading
to Git.