New Server
Created a new EC2 server with an Ubuntu image as opposed to the AWS AMI Linux. Login information for the new server has been added in the README and new key files have been supplied. A dummy servercall c++ file has been added for future testing of the connection between code and the SQL server.