Automatic Range Expansion
Added feature to automatically increase search range if the current area has fewer than the user's specified desired number of POIs

Also, when searching for locations in a line, if it can't find a location in the current area of the line, it will try to search for extra POIs in the previous or future areas. This is to try to avoid the case where the function returned fewer POIs than the user requested.