Updated C++ Server
**Realized my C++ server files were not properly connected to GitHub -
causing changes and syncs not to go through at times.

Added the querying function to the C++ server - calls an example of
Christian's Query

Adding sendMessage functionality. Currently sends the result of the
query to the clients

OffGit:
Changed Unity Scene to incorporate some of the Escality map finding
content. Integrated basic cooridnate mapping and displays the
information from the server.