Clean Up + Refactor
Modified the database query to directly create Place objects rather than storing server results as strings and then parsing the strings to convert them to Place objects.

The old method was used under the assumption that the server query would return one long comma separated string that would need to be parsed. This is no longer the case.